+++ changelog.md
@@ -49,6 +49,7 @@ Internal
 * Move CLI utilities to a new `cli_utils.py`.
 * Move keybinding utilities to a new `key_binding_utils.py`.
 * Move interactive utilities to `interactive_utils.py`.
+* Move special commands out of `main.py`.
 * Modernize orthography of prompt_toolkit filters.
 * Pin all GitHub Actions to hashes.